Track cache updated time more accurately
Changed the 'last_run' key to 'last_updated' to be less confusing. Then,
I made the match counting function responsible for updating it when it
updates the rest of the cache, not the s:IsCacheStale() function. That
makes things a little more accurate and makes a slight optimization
(since earlier a code path wasn't updating the cache update time).