Require +reltime feature
Reltime is quite an old feature and is compiled into any Vim using the
"normal" profile. I feel safe relying on its existence entirely and
simplifying a lot of code this way. If +reltime isn't there, it's likely
enough other things are missing that this plugin doesn't make a great
deal of sense (like +statusline).