Allow counting matches on demand with a command
Adds a command which allows forcing on the match-counting functionality
(if need be), then counting the matches before printing them under the
statusline.

Yet to do are things such as gracefully handling toggling off the
automatic functionality from the configuration; allowing parameters to
be added to count items ad-hoc without a search first; and updating the
documentation.