Ignore most modes to update statusline
A bug report showed that replacing text using the `c` flag, which pauses
for confirmation on each match, broke when this plugin was enabled. This
was because of how the calculations manipulated the search-and-replace
state of the Vim buffer.

We only need to update in the case of the most common modes, and all
others can just wait.

Fixes #1.